backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]
pydantic<3.0.0,>=2.7.4
pydantic-settings>=2.4.0
openai<2.0.0,>=1.40.0
langchain-openai<0.3.0,>=0.2.14
langchain-neo4j==0.4.0
neo4j-driver
python-dotenv>=0.21.0
pandas>=1.0
rag_evaluation==0.1.0

